cosmos/torch_model/requirements.txt
numpy==1.15.4
pyyaml
scikit-image
bitmath
tensorboardX
tensorboard
tqdm
opencv-python
pascal-voc-writer
imgaug
dominate
pytesseract
Pillow
redis
